chore: refactoring card dialog
