+++ src/pages/index.jsx
@@ -2,6 +2,8 @@ import Head from "next/head";
 import React from "react";
 import useGet from "../ions/hooks/fetch/get";
 import Layout from "../organisms/layout";
+import Gallery from "/src/organisms/gallery";
+import Form from "/src/organisms/form";
 
 const Page = () => {
 	const { data, loading, error } = useGet("/api/hello");
@@ -20,6 +22,8 @@ const Page = () => {
 					<code>{JSON.stringify(data, null, 4)}</code>
 				</pre>
 			)}
+			<Gallery />
+			<Form />
 		</Layout>
 	);
 };
